Extract RJiterError/Result to own file.
