+++ rjiter/src/lib.rs
@@ -1,5 +1,8 @@
 pub mod buffer;
+pub mod error;
 pub mod rjiter;
 
 pub use jiter::*;
 pub use rjiter::RJiter;
+pub use error::Error;
+pub use error::Result;
